Code/requirements.txt
opencv-python-headless~=4.10.0.84
numpy~=1.23.0
torch~=2.3.1
torchvision~=0.18.1
pillow~=10.2.0
matplotlib~=3.9.0
requests~=2.32.3
replicate~=0.28.0
mediapipe~=0.10.14